ref(cloudflare): Move D1 instrumentation
