+++ packages/cloudflare/src/index.ts
@@ -127,7 +127,7 @@ export { fetchIntegration } from './integrations/fetch';
 export { vercelAIIntegration } from './integrations/tracing/vercelai';
 export { honoIntegration } from './integrations/hono';
 
-export { instrumentD1WithSentry } from './d1';
+export { instrumentD1WithSentry } from './instrumentations/worker/instrumentD1';
 
 export { instrumentWorkflowWithSentry } from './workflows';
