+++ packages/cloudflare/test/instrumentations/worker/instrumentD1.test.ts
@@ -1,7 +1,7 @@
 import type { D1Database, D1PreparedStatement } from '@cloudflare/workers-types';
 import * as SentryCore from '@sentry/core';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import { instrumentD1WithSentry } from '../src/d1';
+import { instrumentD1WithSentry } from '../../../src/instrumentations/worker/instrumentD1';
 
 const MOCK_FIRST_RETURN_VALUE = { id: 1, name: 'Foo' };
